+++ src/main/java/com/example/sendowl/config/RedisConfigure.java
@@ -1,7 +1,7 @@
 package com.example.sendowl.config;
 
-import com.example.sendowl.redis.RedisMessageSubscriber;
-import com.example.sendowl.redis.RedisShadowkey;
+import com.example.sendowl.redis.sub.RedisMessageSubscriber;
+import com.example.sendowl.redis.service.RedisShadowKeyService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -54,8 +54,8 @@ public class RedisConfigure {
     }
 
     @Bean
-    public RedisShadowkey redisShadowkey(){
-        return new RedisShadowkey();
+    public RedisShadowKeyService redisShadowkey(){
+        return new RedisShadowKeyService();
     }
 
     @Bean // Redis 템플릿을 어디서든 사용할 수 있도록 Bean객체를 생성해준다.
